Collision
Now the main character checks to see if he can move to the space before he does. There is now collision between objects.
If there is a tile on the space you want to move to, it checks to see if the z value is less than that of the main character first before he moves.