Stat Buffing and CPU Decision Making
You are now able to select the move "Intimidate", which lowers the enimies attack stat. A new method, buffStat, takes in the buff along with the stat that you are buffing. All stat buffs disapear after the battle is done.
There is also an animation for taking damage now in the damage indicator.